Break utils.hpp into utils.cpp and utils.hpp
