+++ cpp/benchmarks/CMakeLists.txt
@@ -177,11 +177,11 @@ ConfigureBench(TRANSPOSE_BENCH transpose/transpose.cpp)
 
 # ##################################################################################################
 # * ndsh benchmark --------------------------------------------------------------------------------
-ConfigureNVBench(NDSH_Q1 ndsh/q1.cpp)
-ConfigureNVBench(NDSH_Q5 ndsh/q5.cpp)
-ConfigureNVBench(NDSH_Q6 ndsh/q6.cpp)
-ConfigureNVBench(NDSH_Q9 ndsh/q9.cpp)
-ConfigureNVBench(NDSH_Q10 ndsh/q10.cpp)
+ConfigureNVBench(NDSH_Q1 ndsh/q1.cpp ndsh/utils.cpp)
+ConfigureNVBench(NDSH_Q5 ndsh/q5.cpp ndsh/utils.cpp)
+ConfigureNVBench(NDSH_Q6 ndsh/q6.cpp ndsh/utils.cpp)
+ConfigureNVBench(NDSH_Q9 ndsh/q9.cpp ndsh/utils.cpp)
+ConfigureNVBench(NDSH_Q10 ndsh/q10.cpp ndsh/utils.cpp)
 
 # ##################################################################################################
 # * stream_compaction benchmark -------------------------------------------------------------------
