+++ cpp/benchmarks/ndsh/q1.cpp
@@ -17,6 +17,7 @@
 #include "utils.hpp"
 
 #include <cudf/ast/expressions.hpp>
+#include <cudf/binaryop.hpp>
 #include <cudf/column/column.hpp>
 #include <cudf/scalar/scalar.hpp>
 #include <cudf/utilities/memory_resource.hpp>
