+++ cpp/benchmarks/ndsh/q6.cpp
@@ -17,9 +17,11 @@
 #include "utils.hpp"
 
 #include <cudf/ast/expressions.hpp>
+#include <cudf/binaryop.hpp>
 #include <cudf/column/column.hpp>
 #include <cudf/scalar/scalar.hpp>
 #include <cudf/utilities/memory_resource.hpp>
+#include <cudf/unary.hpp>
 
 #include <nvbench/nvbench.cuh>
